+++ src/main/java/com/onlinebooking/dao/TicketBookingDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.onlinebooking.entities.Ticket;
 
-public interface TicketBookingDao extends CrudRepository<Ticket, Long>{
+public interface TicketBookingDao extends CrudRepository<Ticket, Integer>{
 	Ticket findByEmail(String email);
 }
